Fixing MSVC errors in maintainer build (see issue #901)
git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@11887 21252fe5-60e8-0310-8ff0-c1ae1001594e

+++ 3rd-party/dparser/dparse_tables.h
@@ -137,7 +137,7 @@ typedef struct D_State {
 typedef struct D_Symbol {
   unsigned int		kind;
   const char	        *name;
-  int			name_len;
+  size_t			name_len;
   int			start_symbol;
 } D_Symbol;
